F-1723 F-1724 F-1725 F-2908 F-2910 F-2911 F-3461 F-3683 F-3684 F-3891 F-4605 F-4606 F-4607: fix CAAM, Renesas, SGX, and UEFI examples issues

+++ Renesas/cs+/RH850/rsapss_sign_verify/example_main.c
@@ -91,7 +91,7 @@ size_t  _REL_sizeof_sysheap = SIZEOF_HEAP;
 
 void main(void)
 {
-   byte ret;
+   int ret;
    
    R_SYSTEM_ClockInit();
    R_SYSTEM_TimerInit();
